repropagate changes from repropose-api-ref#210 (#154)
Repropagate changes to object-storage-service-api-ref from PreProd environment

This is a manually created Pull Request for changes to object-storage-service-api-ref to promote changes down the pipeline to Production environment.
It is created in place of #153 because of patch conflicts.
The original change can be found under https://gitea.eco.tsi-dev.otc-service.com/docs/object-storage-service/pulls/210.

Reviewed-by: marmulle2