add changes from propose-api-ref#211 (#161)
Repropagate changes to object-storage-service-api-ref from PreProd environment

This is a manually created Pull Request for changes to object-storage-service-api-ref to promote changes down the pipeline to Production environment.
The original change can be found under https://gitea.eco.tsi-dev.otc-service.com/docs/object-storage-service/pulls/211.
This PR was opened in place of #159.

Reviewed-by: marmulle2