Repropagate changes from https://gitea.eco.tsi-dev.otc-service.com/docs/object-storage-service/pulls/227 (#165)
Repropagate changes to object-storage-service-api-ref from PreProd environment

This is a manually created Pull Request for changes to object-storage-service-api-ref to promote changes down the pipeline to Production environment.
This PR was created to replace #164 because of conflicts.
The original change can be found under https://gitea.eco.tsi-dev.otc-service.com/docs/object-storage-service/pulls/227.

Reviewed-by: marmulle2